feat: better web app flow
Have the client fetch the token from the mod, instead of POSTing the
token to the backend and having it generate HTML to store it in
localStorage.
